2019/12/3/21:31

+++ bos/src/main/java/com/ssm/mapper/SorStorageMapper.java
@@ -18,23 +18,26 @@ public interface SorStorageMapper {
     List<SorStorage> findAllSorStorage();
 
     @Select("select * from (select d.*, ROWNUM ROWNUM_  from (select * from sor_storage) d where ROWNUM <=#{end})where ROWNUM_>#{start} and id like #{id} and acceptdate <= #{acceptdate}")
-    List<SorStorage> selectByPrimaryKeyLike(int start, int end, long id, Date acceptdate);
+    List<SorStorage> selectByPrimaryKeyLike(int start, int end, Integer
+
+
+            id, Date acceptdate);
 
     @Select("select count(*) from sor_storage where id like #{id} and acceptdate <= #{acceptdate}")
-    public int selectCount(Long id, Date acceptdate);
+    public int selectCount(Integer id, Date acceptdate);
 
 
 
 
     @Delete("delete from SOR_Storage where ID=#{id}")
-    int deleteByPrimaryKey(long id);
+    int deleteByPrimaryKey(Integer id);
 
     @Insert("insert into SOR_Storage values(#{id},#{acceptdate},#{acceptperson},#{acceptcompany},#{deliveryperson},#{deliverycompany})")
     int insert(SorStorage record);
 
 
     @Select("select * from SOR_Storage where ID=#{id}")
-    SorStorage selectByPrimaryKey(Short id);
+    SorStorage selectByPrimaryKey(Integer id);
 
     @Update("update  SOR_Storage set acceptdate=#{acceptdate},acceptperson=#{acceptperson},acceptcompany=#{acceptcompany},deliveryperson=#{deliveryperson},deliverycompany=#{deliverycompany} where ID=#{id}")
     int updateByPrimaryKeySelective(SorStorage record);
